@@ -14,15 +14,18 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  platform: '>=2.0.0 <=3.0.0'
-  meta: ^1.0.5
+
+  platform: ^3.0.0-nullsafety.4
+  meta: ^1.3.0-nullsafety.6
+
 dev_dependencies:
-  test: ^1.3.0
-  mockito: ^3.0.0
   flutter_test:
     sdk: flutter
-  pedantic: ^1.9.2
+
+  test: ^1.16.0-nullsafety.17
+  mockito: ^5.0.0-nullsafety.5
+  pedantic: ^1.10.0-nullsafety.3
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.12.0-133.7.beta <3.0.0"
   flutter: ">=1.12.13+hotfix.5 <2.0.0"
