@@ -36,7 +36,7 @@ class AndroidIntent {
     this.arguments,
     this.package,
     this.componentName,
-    Platform platform,
+    Platform? platform,
     this.type,
   })  : assert(action != null || componentName != null,
             'action or component (or both) must be specified'),
@@ -47,8 +47,8 @@ class AndroidIntent {
   /// app code, it may break without warning.
   @visibleForTesting
   AndroidIntent.private({
-    @required Platform platform,
-    @required MethodChannel channel,
+    required Platform platform,
+    required MethodChannel channel,
     this.action,
     this.flags,
     this.category,
@@ -66,47 +66,47 @@ class AndroidIntent {
   /// includes constants like `ACTION_VIEW`.
   ///
   /// See https://developer.android.com/reference/android/content/Intent.html#intent-structure.
-  final String action;
+  final String? action;
 
   /// Constants that can be set on an intent to tweak how it is finally handled.
   /// Some of the constants are mirrored to Dart via [Flag].
   ///
   /// See https://developer.android.com/reference/android/content/Intent.html#setFlags(int).
-  final List<int> flags;
+  final List<int>? flags;
 
   /// An optional additional constant qualifying the given [action].
   ///
   /// See https://developer.android.com/reference/android/content/Intent.html#intent-structure.
-  final String category;
+  final String? category;
 
   /// The Uri that the [action] is pointed towards.
   ///
   /// See https://developer.android.com/reference/android/content/Intent.html#intent-structure.
-  final String data;
+  final String? data;
 
   /// The equivalent of `extras`, a generic `Bundle` of data that the Intent can
   /// carry. This is a slot for extraneous data that the listener may use.
   ///
   /// See https://developer.android.com/reference/android/content/Intent.html#intent-structure.
-  final Map<String, dynamic> arguments;
+  final Map<String, dynamic>? arguments;
 
   /// Sets the [data] to only resolve within this given package.
   ///
   /// See https://developer.android.com/reference/android/content/Intent.html#setPackage(java.lang.String).
-  final String package;
+  final String? package;
 
   /// Set the exact `ComponentName` that should handle the intent. If this is
   /// set [package] should also be non-null.
   ///
   /// See https://developer.android.com/reference/android/content/Intent.html#setComponent(android.content.ComponentName).
-  final String componentName;
+  final String? componentName;
   final MethodChannel _channel;
   final Platform _platform;
 
   /// Set an explicit MIME data type.
   ///
   /// See https://developer.android.com/reference/android/content/Intent.html#intent-structure.
-  final String type;
+  final String? type;
 
   bool _isPowerOfTwo(int x) {
     /* First x in the below expression is for the case when x is 0 */
@@ -176,17 +176,17 @@ class AndroidIntent {
       return false;
     }
 
-    return await _channel.invokeMethod<bool>(
+    return (await _channel.invokeMethod<bool>(
       'canResolveActivity',
       _buildArguments(),
-    );
+    ))!;
   }
 
   /// Constructs the map of arguments which is passed to the plugin.
   Map<String, dynamic> _buildArguments() {
     return {
       if (action != null) 'action': action,
-      if (flags != null) 'flags': convertFlags(flags),
+      if (flags != null) 'flags': convertFlags(flags!),
       if (category != null) 'category': category,
       if (data != null) 'data': data,
       if (arguments != null) 'arguments': arguments,
